small adjustment

+++ YOLOv8-Instance-Segmentation-Android-Tflite/app/src/main/java/com/surendramaran/yolov8_instancesegmentation/ml/SegmentationResult.kt
@@ -2,5 +2,5 @@ package com.surendramaran.yolov8_instancesegmentation.ml
 
 data class SegmentationResult(
     val box: Output0,
-    val mask: List<DoubleArray>
+    val mask: List<IntArray>
 )
